% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_delete_members}
\alias{guardduty_delete_members}
\title{Deletes GuardDuty member accounts (to the current GuardDuty master
account) specified by the account IDs}
\usage{
guardduty_delete_members(AccountIds, DetectorId)
}
\arguments{
\item{AccountIds}{[required] A list of account IDs of the GuardDuty member accounts that you want to delete.}

\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty account whose members you want to delete.}
}
\description{
Deletes GuardDuty member accounts (to the current GuardDuty master account) specified by the account IDs.
}
\section{Request syntax}{
\preformatted{guardduty$delete_members(
  AccountIds = list(
    "string"
  ),
  DetectorId = "string"
)
}
}

\keyword{internal}
