% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_get_master_account}
\alias{guardduty_get_master_account}
\title{Provides the details for the GuardDuty master account to the current
GuardDuty member account}
\usage{
guardduty_get_master_account(DetectorId)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector of the GuardDuty member account.}
}
\description{
Provides the details for the GuardDuty master account to the current GuardDuty member account.
}
\section{Request syntax}{
\preformatted{guardduty$get_master_account(
  DetectorId = "string"
)
}
}

\keyword{internal}
