% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_list_ip_sets}
\alias{guardduty_list_ip_sets}
\title{Lists the IPSets of the GuardDuty service specified by the detector ID}
\usage{
guardduty_list_ip_sets(DetectorId, MaxResults, NextToken)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to retrieve.}

\item{MaxResults}{You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 7. The maximum value is 7.}

\item{NextToken}{You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListIPSet action. For subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.}
}
\description{
Lists the IPSets of the GuardDuty service specified by the detector ID.
}
\section{Request syntax}{
\preformatted{guardduty$list_ip_sets(
  DetectorId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

\keyword{internal}
