% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guardduty_operations.R
\name{guardduty_update_detector}
\alias{guardduty_update_detector}
\title{Updates an Amazon GuardDuty detector specified by the detectorId}
\usage{
guardduty_update_detector(DetectorId, Enable,
  FindingPublishingFrequency)
}
\arguments{
\item{DetectorId}{[required] The unique ID of the detector that you want to update.}

\item{Enable}{Updated boolean value for the detector that specifies whether the detector is enabled.}

\item{FindingPublishingFrequency}{A enum value that specifies how frequently customer got Finding updates published.}
}
\description{
Updates an Amazon GuardDuty detector specified by the detectorId.
}
\section{Request syntax}{
\preformatted{guardduty$update_detector(
  DetectorId = "string",
  Enable = TRUE|FALSE,
  FindingPublishingFrequency = "FIFTEEN_MINUTES"|"ONE_HOUR"|"SIX_HOURS"
)
}
}

\keyword{internal}
